added similarityTab, all measures are now augmented correctly, not all bugs fixed

+++ src/walkerPack/SetSequence.java
@@ -18,7 +18,7 @@ public class SetSequence
     // 9 nucleotides:  G A U C R M S V N
     // store them in array , with index starting from 0 as specified
 
-    public int nucleotides[];
+    public double nucleotides[];
 
     public SetSequence( String sequence )
     {
